move unit tests to utests/test_utils.py
