+++ utests/test_utils.py
@@ -4,9 +4,15 @@ import pytest
 from requests import Session
 
 from RequestsLibrary import RequestsLibrary
-from RequestsLibrary.utils import is_file_descriptor, merge_headers
+from RequestsLibrary.utils import is_file_descriptor, merge_headers, process_secrets
 from utests import SCRIPT_DIR
 from utests import mock
+
+try:
+    from robot.api.types import Secret
+    secret_type_supported = True
+except (ImportError, ModuleNotFoundError):
+    secret_type_supported = False
 
 
 def test_none():
@@ -72,3 +78,29 @@ def test_warn_that_url_is_missing(mocked_logger, mocked_keywords):
     except TypeError:
         pass
     mocked_logger.warn.assert_called()
+
+
+def test_process_secrets_with_no_secrets():
+    auth = ('user', 'password')
+    result = process_secrets(auth)
+    assert result == ('user', 'password')
+
+
+@pytest.mark.skipif(not secret_type_supported, reason="Running on pre-7.4 robot")
+def test_process_secrets_with_secrets():
+    secret_password = Secret('mypassword')
+    auth = ('user', secret_password)
+    result = process_secrets(auth)
+    assert result == ('user', 'mypassword')
+    assert not isinstance(result[1], Secret)
+
+
+@pytest.mark.skipif(not secret_type_supported, reason="Running on pre-7.4 robot")
+def test_process_secrets_with_mixed_secrets():
+    secret_user = Secret('myuser')
+    secret_password = Secret('mypassword')
+    auth = (secret_user, secret_password)
+    result = process_secrets(auth)
+    assert result == ('myuser', 'mypassword')
+    assert not isinstance(result[0], Secret)
+    assert not isinstance(result[1], Secret)
